pyproject.toml
[tool.poetry]
name = "configsys"
version = "0.1.0"
description = ""
authors = ["Erik Gafni <egafni@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
dacite = "^1.8.1"
fsspec = "^2023.10.0"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
